requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
requests
